worked on the Unit 7: Assignment for Game dev
Worked on trying to make a ai chaser, I had to do a lotttt of tweaking and stuff for this one, I reimported the artisan tools and did a bunch of other stuff because it just was not working.

So if something breaks later on, this might be the culprit.